+++ src/frontend/src/pages/NewCalendarPage/EventPartialInfoView.tsx
@@ -3,7 +3,7 @@ import LocationOnIcon from '@mui/icons-material/LocationOn';
 import { Calendar, DayOfWeek, Event, EventType } from 'shared';
 import GroupIcon from '@mui/icons-material/Group';
 import { Stack } from '@mui/system';
-import { getTeamTypeIcon } from './EventClickPopup';
+import { getTeamTypeIcon } from './CalendarDayCard';
 import { Typography } from '@mui/material';
 import { getConvertedEnd, getConvertedStart } from '../../utils/datetime.utils';
